+++ src/bit-systems/media-loading.ts
@@ -6,8 +6,8 @@ import { ErrorObject } from "../prefabs/error-object";
 import { LoadingObject } from "../prefabs/loading-object";
 import { animate } from "../utils/animate";
 import { setNetworkedDataWithoutRoot } from "../utils/assign-network-ids";
-import { cancelable, coroutine, crClearTimeout, crNextFrame, crTimeout, makeCancelable } from "../utils/coroutine";
-import { Job, JobMap, startJob, stopJob, tickJobs } from "../utils/coroutine-utils";
+import { coroutine, crClearTimeout, crNextFrame, crTimeout, makeCancelable } from "../utils/coroutine";
+import { cancelable, Job, JobMap, startJob, stopJob, tickJobs } from "../utils/coroutine-utils";
 import { easeOutQuadratic } from "../utils/easing";
 import { renderAsEntity } from "../utils/jsx-entity";
 import { loadImage } from "../utils/load-image";
