+++ src/bit-systems/pdf-system.ts
@@ -4,8 +4,8 @@ import { Mesh, MeshBasicMaterial } from "three";
 import { HubsWorld } from "../app";
 import { MediaPDF, NetworkedPDF } from "../bit-components";
 import { PDFComponent } from "../inflators/pdf";
-import { cancelable, coroutine, makeCancelable } from "../utils/coroutine";
-import { JobMap, startJob, stopJob, tickJobs } from "../utils/coroutine-utils";
+import { coroutine, makeCancelable } from "../utils/coroutine";
+import { cancelable, JobMap, startJob, stopJob, tickJobs } from "../utils/coroutine-utils";
 import { EntityID } from "../utils/networking-types";
 import { scaleMeshToAspectRatio } from "../utils/scale-to-aspect-ratio";
 import { waitForMediaLoaded } from "./media-loading";
